safari-534.53-branch/LayoutTests/fast/inspector-support/uncaught-dom8-exception-expected.txt

CONSOLE MESSAGE: line 9: NOT_FOUND_ERR: DOM Exception 8: An attempt was made to reference a Node in a context where it does not exist.
This tests that we are getting the correct message for DOM Exception 8: NOT_FOUND_ERR.
